Fix tests structure
